+++ XLineTools/XCSourceTextBuffer+SourceTextBuffer.swift
@@ -28,6 +28,10 @@ extension XCSourceTextBuffer: SourceTextBuffer {
         lines[index] as? String
     }
 
+    func replaceLine(at index: Int, with line: String) {
+        lines[index] = line
+    }
+
     func insert(line: String, at index: Int) {
         lines.insert(line, at: index)
     }